Перенес файлы в папку Lab3
