+++ Lab3/Iterator.py
